Add assign-scalar for sparse vector
